Add comments. Move Form to common folder
